Arreglos visuales cliente y host. No hace mas falta agregar el IPV4 en config, lo intenta obtener automaticamente.

+++ src/main/java/com/sabu/utils/Translate.java
@@ -2,12 +2,12 @@ package com.sabu.utils;
 
 public class Translate {
 
-    public static char translateCharToNumber(String toTransform){
+    public static char translateCharToNumber(String toTransform) {
         return (char) (toTransform.charAt(0) + 17);
     }
 
-    public static int translateIntToChar(char toTransform,int value){
-        return (int) (toTransform - value);
+    public static int translateIntToChar(char toTransform, int value) {
+        return toTransform - value;
     }
